// src/radio-v2/types.ts
export type LabelPlacement = keyof typeof LABEL_PLACEMENT;
export type Align = keyof typeof ALIGN;

export type RadioOverrides = {
  RadioBackplate?: Override;
  RadioMarkInner?: Override;
  RadioMarkOuter?: Override;
  Label?: Override;
  LabelWrapper?: Override;
  Root?: Override;
  Input?: Override;
  Description?: Override;
};
// ...
export type RadioProps = {
  /** Id of element which contains a related caption */
  'aria-describedby'?: string;
  /** Id of element which contains a related error message */
  'aria-errormessage'?: string;
  /**
   * Used to define a string that labels the radio. Use this prop if the label is not
   * visible on screen. If the label is visible, use the 'aria-labeledby' prop instead.
   */
  'aria-label'?: string;
  /**
   * Establishes a relationship between the radio and its label. Screen readers use this
   * attribute to catalog the object on a page so that users can navigate between them.
   */
  'aria-labelledby'?: string;
  /** Focus the radio on initial render. */
  autoFocus?: boolean;
  /** How the radio will be displayed along with its description. Controls spacing */
  align?: Align;
  /** Check or uncheck the control. */
  checked?: boolean;
  /** Label of radio. */
  children?: React.ReactNode;
  /** Indicates if this radio children contain an interactive element (prevents the label from moving focus from the child element to the radio button) */
  containsInteractiveElement?: boolean;
  /** Add more detail about a radio element. */
  description?: string;
  /** Disable the checkbox from being changed. */
  disabled?: boolean;
  /** Used to get a ref to the input element. Useful for programmatically focusing the input */
  inputRef?: React.RefObject<HTMLInputElement>;
  /** Renders checkbox in errored state. */
  error?: boolean;
  /** Is radio focused / active? */
  isFocused?: boolean;
  /** Is parent RadioGroup focused by keyboard? */
  isFocusVisible?: boolean;
  /** How to position the label relative to the checkbox itself. */
  labelPlacement?: LabelPlacement;
  /** Passed to the input element name attribute */
  name?: string;
  /** handler for blur events on trigger element. */
  onBlur?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** Handler for change events on trigger element. */
  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** handler for focus events on trigger element. */
  onFocus?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** Handler for mouseenter events on trigger element. */
  onMouseEnter?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** Handler for mouseleave events on trigger element. */
  onMouseLeave?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** Handler for mousedown events on trigger element. */
  onMouseDown?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  /** Handler for mouseup events on trigger element. */
  onMouseUp?: (e: React.ChangeEvent<HTMLInputElement>) => unknown;
  overrides?: RadioOverrides;
  /** Marks the checkbox as required. */
  required?: boolean;
  /** Passed to the input element value attribute */
  value?: string;
  /** Passed to the input element, typically managed by RadioGroup */
  tabIndex?: string;
};